+++ packages/prettier/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @minko-fe/prettier-config
 
+## 2.2.0
+
+### Minor Changes
+
+- feat: suppor `astro` and `svelte` by default
+
 ## 2.1.0
 
 ### Minor Changes